apps/backend/pyproject.toml
[project]
name = "finance-foundation-backend"
version = "0.1.0"
description = "FastAPI backend for finance-foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "email-validator>=2.2.0",
  "httpx>=0.28.0",
  "argon2-cffi>=23.1.0",
  "itsdangerous>=2.2.0",
  "pyjwt[crypto]>=2.10.0",
  "pwdlib>=0.2.1",
  "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
  "pyright>=1.1.403",
  "pre-commit>=4.3.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501"]

[tool.pyright]
include = ["app", "tests"]
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
